feat: extract API endpoint modal to Cluster plugin
